* update to bullet 2.86.1 https://github.com/bulletphysics/bullet3/releases/tag/2.86.1
fix all swig warnings
update jni header files
add BulletInverseDynamics
add Vehicle

For issue: https://github.com/libgdx/libgdx/issues/4614

+++ extensions/gdx-bullet/jni/src/bullet/BulletDynamics/Featherstone/btMultiBodyDynamicsWorld.h
@@ -58,7 +58,7 @@ public:
 
 	virtual ~btMultiBodyDynamicsWorld ();
 
-	virtual void	addMultiBody(btMultiBody* body, short group= btBroadphaseProxy::DefaultFilter, short mask=btBroadphaseProxy::AllFilter);
+	virtual void	addMultiBody(btMultiBody* body, int group= btBroadphaseProxy::DefaultFilter, int mask=btBroadphaseProxy::AllFilter);
 
 	virtual void	removeMultiBody(btMultiBody* body);
